build: update build process
