+++ .npmrc
@@ -1,1 +1,2 @@
 message="chore(release): %s :tada:"
+registry='https://registry.npmjs.org'
