+++ packages/oruga-next/package.json
@@ -21,7 +21,8 @@
     "dist",
     "src",
     "nuxt",
-    "types/*.d.ts"
+    "types/*.d.ts",
+    "README.md"
   ],
   "sideEffects": [
     "*.css",
@@ -37,17 +38,13 @@
   "scripts": {
     "build:vue": "rollup -c && rollup -c --environment MINIFY",
     "build:vue:watch": "rollup -c --watch",
-    "build:style": "npm run build:scss && npm run build:scss-vars && npm run build:scss-full",
-    "build:scss-vars": "sass --no-charset ../oruga/src/scss/oruga-full-vars.scss | node ../oruga/build/banner.js > dist/oruga-full-vars.css && cleancss -o dist/oruga-full-vars.min.css dist/oruga-full-vars.css",
-    "build:scss-full": "sass --no-charset ../oruga/src/scss/oruga-full.scss | node ../oruga/build/banner.js > dist/oruga-full.css && cleancss -o dist/oruga-full.min.css dist/oruga-full.css",
-    "build:scss": "sass --no-charset ../oruga/src/scss/oruga.scss | node ../oruga/build/banner.js > dist/oruga.css && cleancss -o dist/oruga.min.css dist/oruga.css",
-    "build:style:watch": "concurrently \"sass ../oruga/src/scss/oruga-full-vars.scss dist/oruga-full-vars.css --watch\" \"sass ../oruga/src/scss/oruga.scss dist/oruga.css --watch\"",
-    "build:lib": "rimraf dist && npm run build:vue && npm run build:style",
-    "build:lib:watch": "npm link && concurrently \"npm run build:vue:watch\" \"npm run build:style:watch\"",
-    "publish:lib": "cp -R ../oruga/src/scss ./src && cp ../../README.md . && npm run build:lib && npm publish",
+    "build:lib": "rimraf dist && npm run build:vue",
+    "build:lib:watch": "npm link &&  \"npm run build:vue:watch\"",
+    "publish:lib": "cp ../../README.md . && npm run build:lib && npm publish",
     "test": "rm -rf .nyc_output coverage && NODE_ENV=test cypress run --component",
     "test:ts": "vue-tsc --noEmit --skipLibCheck",
-    "test:watch": "rm -rf .nyc_output coverage && NODE_ENV=test cypress open --component"
+    "test:watch": "rm -rf .nyc_output coverage && NODE_ENV=test cypress open --component",
+    "update": "ncu -u"
   },
   "keywords": [
     "oruga",
@@ -64,46 +61,54 @@
     "vue": "^3.0.0"
   },
   "devDependencies": {
-    "@babel/core": "7.10.2",
-    "@babel/preset-env": "7.10.2",
-    "@cypress/code-coverage": "^3.10.0",
-    "@cypress/vite-dev-server": "^3.4.0",
-    "@cypress/vue": "^4.2.2",
-    "@rollup/plugin-babel": "5.0.2",
-    "@rollup/plugin-commonjs": "12.0.0",
-    "@rollup/plugin-node-resolve": "8.0.0",
-    "@rollup/plugin-replace": "2.3.2",
-    "@rollup/plugin-typescript": "6.1.0",
-    "@types/jest": "24.0.19",
-    "@types/node": "^18.11.12",
-    "@typescript-eslint/eslint-plugin": "2.26.0",
-    "@typescript-eslint/parser": "2.26.0",
-    "@vitejs/plugin-vue": "^3.2.0",
-    "@vue/compiler-sfc": "^3.2.45",
-    "@vue/eslint-config-typescript": "^11.0.2",
+    "@babel/core": "7.22.17",
+    "@babel/preset-env": "7.22.15",
+    "@cypress/code-coverage": "^3.12.1",
+    "@cypress/vite-dev-server": "^5.0.6",
+    "@cypress/vue": "^6.0.0",
+    "@rollup/plugin-babel": "6.0.3",
+    "@rollup/plugin-commonjs": "25.0.4",
+    "@rollup/plugin-node-resolve": "15.2.1",
+    "@rollup/plugin-replace": "5.0.2",
+    "@rollup/plugin-terser": "^0.4.3",
+    "@rollup/plugin-typescript": "11.1.3",
+    "@types/jest": "29.5.4",
+    "@types/node": "^20.6.0",
+    "@vitejs/plugin-vue": "^4.3.4",
+    "@vue/compiler-sfc": "^3.3.4",
+    "@vue/eslint-config-prettier": "^8.0.0",
+    "@vue/eslint-config-standard": "^8.0.1",
+    "@vue/eslint-config-typescript": "^12.0.0",
+    "@vue/server-renderer": "^3.3.4",
+    "@vue/test-utils": "^2.4.1",
+    "@vue/tsconfig": "^0.4.0",
     "babel-plugin-istanbul": "^6.1.1",
-    "clean-css-cli": "4.3.0",
-    "codecov": "3.7.2",
-    "concurrently": "5.2.0",
-    "core-js": "3.6.5",
-    "cypress": "^10.11.0",
-    "eslint": "6.7.2",
-    "eslint-plugin-vue": "^9.8.0",
-    "jest-serializer-vue": "0.3.0",
-    "rollup": "^2.60.0",
+    "codecov": "3.8.2",
+    "core-js": "3.32.2",
+    "cypress": "^13.2.0",
+    "deepmerge": "^4.3.1",
+    "eslint": "^8.49.0",
+    "eslint-config-prettier": "^9.0.0",
+    "eslint-config-standard": "^17.1.0",
+    "eslint-config-standard-with-typescript": "^39.0.0",
+    "eslint-plugin-node": "^11.1.0",
+    "eslint-plugin-promise": "^6.1.1",
+    "eslint-plugin-vue": "^9.17.0",
+    "jest-serializer-vue": "3.1.0",
+    "npm-check-updates": "^16.14.2",
+    "prettier": "^3.0.3",
+    "rollup": "^3.29.1",
     "rollup-copy-plugin": "0.1.0",
     "rollup-plugin-eslint": "7.0.0",
-    "rollup-plugin-terser": "6.1.0",
-    "rollup-plugin-typescript2": "^0.34.1",
+    "rollup-plugin-typescript2": "^0.35.0",
     "rollup-plugin-vue": "6.0.0",
-    "sass": "1.27.0",
-    "sinon": "9.0.3",
-    "typescript": "^4.9.3",
-    "vite": "^3.2.2",
-    "vite-plugin-istanbul": "^3.0.1",
-    "vue": "^3.2.45",
-    "vue-tsc": "^1.0.12",
-    "webpack-dev-server": "^4.7.4"
+    "sass": "1.67.0",
+    "sinon": "16.0.0",
+    "typescript": "^5.2.2",
+    "vite": "^4.4.9",
+    "vite-plugin-istanbul": "^5.0.0",
+    "vue": "^3.3.4",
+    "vue-tsc": "^1.8.11"
   },
   "nyc": {
     "all": true,
