+++ packages/oruga-next/src/components/datepicker/DatepickerTableRow.vue
@@ -42,7 +42,7 @@
 </template>
 
 <script lang="ts">
-import { defineComponent, PropType } from 'vue'
+import { defineComponent, type PropType } from 'vue'
 
 import BaseComponentMixin from '../../utils/BaseComponentMixin'
