+++ packages/oruga-next/src/shims-vue.d.ts
@@ -1,5 +1,10 @@
+// shims-vue.d.ts
 declare module '*.vue' {
   import type { DefineComponent } from 'vue'
-  const component: DefineComponent<{}, {}, any>
+  const component: DefineComponent<
+    Record<string, unknown>,
+    Record<string, unknown>,
+    any
+  >; 
   export default component
 }
